[Feat] 지도 친구선택 유저 수 대응
